+++ pom.xml
@@ -234,14 +234,12 @@
         </dependency>
 
         <!-- Testing -->
-        <!-- Debezium Scripting for Filter Transform -->
         <dependency>
             <groupId>io.debezium</groupId>
             <artifactId>debezium-scripting</artifactId>
             <version>${version.debezium}</version>
             <scope>test</scope>
         </dependency>
-        <!-- Groovy for JSR 223 scripting engine -->
         <dependency>
             <groupId>org.codehaus.groovy</groupId>
             <artifactId>groovy</artifactId>
