+++ src/main/java/io/debezium/connector/yugabytedb/YugabyteDBConnectorTask.java
@@ -464,6 +464,7 @@ public class YugabyteDBConnectorTask
                                              taskContext.getTaskId(), entry.getValue(), entry.getKey());
                             }
                         }
+                        
                         if (ybOffset != null) {
                             this.coordinator.commitOffset(ybOffset);
                         }
